Improve tests & update version to 1.3.1 (#166)
- Tests now ignore the version change (to avoid regenerating a bunch of golden files)
- Fix shader compilation warning (int vs float)

+++ tests/unit/inline-aggro.frag
@@ -119,7 +119,7 @@ int noinl11(in ivec3 x) {
 	return x[i] + i;
 }
 
-float noinl12() {
+int noinl12() {
 	int i = 10;
 	while (--i > 0) {
 	}
